impr: Move Inputs to it own plugin
